SDI-603: ✨ Tidy up tests and only index centimetres
