+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prisonersearch/resource/PrisonerDetailResourceTest.kt
@@ -1,16 +1,18 @@
 package uk.gov.justice.digital.hmpps.prisonersearch.resource
 
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.web.reactive.function.BodyInserters
 import uk.gov.justice.digital.hmpps.prisonersearch.QueueIntegrationTest
+import uk.gov.justice.digital.hmpps.prisonersearch.model.RestResponsePage
 import uk.gov.justice.digital.hmpps.prisonersearch.services.dto.KeywordRequest
 import uk.gov.justice.digital.hmpps.prisonersearch.services.dto.PrisonerDetailRequest
 
 class PrisonerDetailResourceTest : QueueIntegrationTest() {
 
   companion object {
-    var initialiseSearchData = true
+    private var initialiseSearchData = true
   }
 
   @BeforeEach
@@ -98,7 +100,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by whole prisoner number`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(nomsNumber = "A7089EY", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -107,7 +108,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by whole lowercase prisoner number `() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(nomsNumber = "a7089Ey", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -116,7 +116,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by prisoner number with a wildcard single letter`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(nomsNumber = "A7089?Y", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -125,7 +124,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by prisoner number with wildcard suffix`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(nomsNumber = "A7089*", prisonIds = listOf("MDI")),
-      expectedCount = 3,
       expectedPrisoners = listOf("A7089EY", "A7089FA", "A7089FB"),
     )
   }
@@ -134,7 +132,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by whole PNC number with short year`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(pncNumber = "12/394773H", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -143,7 +140,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by whole PNC number with long year`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(pncNumber = "2015/1234S", prisonIds = listOf("WSI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A9999AA"),
     )
   }
@@ -152,7 +148,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by lowercase PNC number with short year`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(pncNumber = "12/394773h", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -161,7 +156,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by lowercase PNC with long year`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(pncNumber = "2012/394773h", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -170,7 +164,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by PNC number with wildcard single digit`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(pncNumber = "12/39477?H", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -179,7 +172,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by PNC number with a wildcard suffix and matching surname`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(pncNumber = "12/394773*", lastName = "smith", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -188,7 +180,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by whole CRO number`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(croNumber = "29906/12J", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -197,7 +188,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by lowercase CRO number `() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(croNumber = "29906/12j", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -206,7 +196,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by CRO number with wildcard single letter`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(croNumber = "29906/1?J", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY")
     )
   }
@@ -215,7 +204,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by CRO number with wildcard suffix`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(croNumber = "29906/*J", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -224,7 +212,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by criteria that do not match any prisoners - empty result`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "trevor", pncNumber = "29906/12J", prisonIds = listOf("MDI")),
-      expectedCount = 0,
       expectedPrisoners = emptyList(),
     )
   }
@@ -233,8 +220,7 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first name`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "john", prisonIds = listOf("LEI", "MDI")),
-      expectedCount = 2,
-      expectedPrisoners = listOf("A7089EZ", "A7089EY"),
+      expectedPrisoners = listOf("A7089EY", "A7089EZ"),
     )
   }
 
@@ -242,7 +228,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by last name`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(lastName = "smith", prisonIds = listOf("MDI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7089EY"),
     )
   }
@@ -251,7 +236,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first name including aliases`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "Sam", prisonIds = listOf("LEI", "MDI")),
-      expectedCount = 4,
       expectedPrisoners = listOf("A7090AB", "A7090BA", "A7090BB", "A7090AF"),
     )
   }
@@ -260,7 +244,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first name excluding aliases`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "Sam", prisonIds = listOf("LEI", "MDI"), includeAliases = false),
-      expectedCount = 3,
       expectedPrisoners = listOf("A7090AB", "A7090BA", "A7090BB"),
     )
   }
@@ -269,8 +252,7 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by last name including aliases`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(lastName = "Jones", prisonIds = listOf("MDI", "LEI")),
-      expectedCount = 6,
-      expectedPrisoners = listOf("A7090AA", "A7090AB", "A7090BA", "A7090BB", "A7090AF", "A1090AA"),
+      expectedPrisoners = listOf("A1090AA", "A7090AA", "A7090AB", "A7090BA", "A7090BB", "A7090AF"),
     )
   }
 
@@ -278,8 +260,7 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by last name excluding aliases`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(lastName = "Jones", prisonIds = listOf("MDI", "LEI"), includeAliases = false),
-      expectedCount = 5,
-      expectedPrisoners = listOf("A7090AA", "A7090AB", "A7090BA", "A7090BB", "A1090AA"),
+      expectedPrisoners = listOf("A1090AA", "A7090AA", "A7090AB", "A7090BA", "A7090BB"),
     )
   }
 
@@ -287,7 +268,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first and last names`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "sam", lastName = "jones", prisonIds = listOf("MDI", "AGI", "LEI")),
-      expectedCount = 7,
       expectedPrisoners = listOf("A7090AB", "A7090AC", "A7090AD", "A7090BA", "A7090BB", "A7090BC", "A7090AF"),
     )
   }
@@ -296,7 +276,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first and last names excluding aliases`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "sam", lastName = "jones", prisonIds = listOf("MDI", "AGI", "LEI"), includeAliases = false),
-      expectedCount = 6,
       expectedPrisoners = listOf("A7090AB", "A7090AC", "A7090AD", "A7090BA", "A7090BB", "A7090BC"),
     )
   }
@@ -305,7 +284,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by mixed case first and last names`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "Sam", lastName = "Jones", prisonIds = listOf("MDI", "AGI", "LEI")),
-      expectedCount = 7,
       expectedPrisoners = listOf("A7090AB", "A7090AC", "A7090AD", "A7090BA", "A7090BB", "A7090BC", "A7090AF"),
     )
   }
@@ -314,7 +292,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first and last names in alias`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "danny", lastName = "colin", prisonIds = listOf("LEI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7090AF"),
     )
   }
@@ -323,7 +300,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by mixed case first and last names in alias`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "DANny", lastName = "COLin", prisonIds = listOf("LEI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7090AF"),
     )
   }
@@ -332,7 +308,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by first and last names in alias with wildcard letters`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "dann?", lastName = "col?n", prisonIds = listOf("LEI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7090AF"),
     )
   }
@@ -341,7 +316,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by main first and last name with single wildcard letters`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "jimb?b", lastName = "j?cks", prisonIds = listOf("LEI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7090AF"),
     )
   }
@@ -350,7 +324,6 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `find by mixed case first and last name with single wildcard letters`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(firstName = "JIMb?b", lastName = "j?cKs", prisonIds = listOf("LEI")),
-      expectedCount = 1,
       expectedPrisoners = listOf("A7090AF"),
     )
   }
@@ -359,8 +332,25 @@ class PrisonerDetailResourceTest : QueueIntegrationTest() {
   fun `no-terms query should match all prisoners in the specified location`() {
     detailSearch(
       detailRequest = PrisonerDetailRequest(prisonIds = listOf("MDI")),
-      expectedCount = 7,
-      expectedPrisoners = listOf("A7089EY", "A7089FA", "A7089FB", "A7090AA", "A7090AB", "A7090BB", "A1090AA"),
+      expectedPrisoners = listOf("A1090AA", "A7089EY", "A7089FA", "A7089FB", "A7090AA", "A7090AB", "A7090BB"),
     )
+  }
+
+  private fun detailSearch(
+    detailRequest: PrisonerDetailRequest,
+    expectedPrisoners: List<String> = emptyList(),
+  ) {
+    val response = webTestClient.post().uri("/prisoner-detail")
+      .body(BodyInserters.fromValue(gson.toJson(detailRequest)))
+      .headers(setAuthorisation(roles = listOf("ROLE_GLOBAL_SEARCH")))
+      .header("Content-Type", "application/json")
+      .exchange()
+      .expectStatus().isOk
+      .expectBody(RestResponsePage::class.java)
+      .returnResult().responseBody
+
+    assertThat(response.content).extracting("prisonerNumber").containsExactlyElementsOf(expectedPrisoners)
+    assertThat(response.content).size().isEqualTo(expectedPrisoners.size)
+    assertThat(response.numberOfElements).isEqualTo(expectedPrisoners.size)
   }
 }
