+++ src/test/resources/mappings/getBooking_A1090AA.json
@@ -27,7 +27,11 @@
         "gender": "Male",
         "raceCode": "W1",
         "ethnicity": "White: Eng./Welsh/Scot./N.Irish/British",
-        "sexCode": "M"
+        "sexCode": "M",
+        "heightFeet": 6,
+        "heightInches": 8,
+        "heightMetres": 2.02,
+        "heightCentimetres": 202
       },
       "physicalCharacteristics": [
         {
